Split VLLM unit tests for accuracy
