+++ tests/test_cutlass_scaled_mm.py
@@ -1,0 +1,228 @@
+import random
+from itertools import product
+from math import ceil
+from typing import Optional
+
+import pytest
+import torch
+
+import flag_gems
+
+from .conftest import QUICK_MODE
+
+random.seed(42)
+
+try:
+    import vllm  # noqa: 401
+
+    VLLM_AVAILABLE = True
+except ImportError:
+    VLLM_AVAILABLE = False
+
+
+def is_cuda_available():
+    if flag_gems.device != "cuda":
+        return False
+    major, minor = torch.cuda.get_device_capability()
+    sm_version_num = major * 10 + minor
+    return sm_version_num >= 90 and sm_version_num < 100
+
+
+CUDA_AVAILABLE = is_cuda_available()
+
+
+def to_int8(tensor: torch.Tensor):
+    return torch.round(tensor.clamp(min=-128, max=127)).to(dtype=torch.int8)
+
+
+def to_fp8(tensor: torch.Tensor):
+    finfo = torch.finfo(torch.float8_e4m3fn)
+    return torch.round(tensor.clamp(min=finfo.min, max=finfo.max)).to(
+        dtype=torch.float8_e4m3fn
+    )
+
+
+class CutlassScaledMMTestKit:
+    num_test_cases = 16 if QUICK_MODE else 32
+
+    @staticmethod
+    def _get_all_combinations():
+        # these shapes come from the test file of op `cutlass_scaled_mm` of vLLM
+        mnk = [
+            (1, 256, 128),
+            (1, 16384, 1024),
+            (1, 24576, 496),
+            (16, 256, 496),
+            (16, 16384, 128),
+            (16, 24576, 4096),
+            (32, 8192, 4096),
+            (32, 16384, 4096),
+            (33, 1024, 1024),
+            (33, 8192, 128),
+            (64, 2048, 496),
+            (64, 16384, 1024),
+            (100, 8192, 496),
+            (128, 32768, 4096),
+            (256, 4096, 4096),
+            (512, 256, 1024),
+            (512, 8192, 4096),
+            (512, 16384, 128),
+            (512, 24576, 128),
+        ]
+        scale_shape_types = ["scalar", "vector", "matrix"]
+        if_use_bias = [True, False]
+        dtypes = [(torch.int8, torch.float16), (torch.float8_e4m3fn, torch.bfloat16)]
+
+        combinations = product(
+            mnk, scale_shape_types, scale_shape_types, if_use_bias, dtypes
+        )
+        return combinations
+
+    @classmethod
+    def _rand_sample(cls, all_params):
+        random.shuffle(all_params)
+        return all_params[: cls.num_test_cases]
+
+    @classmethod
+    def get_test_params(cls):
+        combinations = cls._get_all_combinations()
+
+        all_params = []
+        for (
+            (M, N, K),
+            a_scale_category,
+            b_scale_category,
+            bias,
+            (in_dtype, out_dtype),
+        ) in combinations:
+            is_scalar_or_vector_dequant = a_scale_category in [
+                "scalar",
+                "vector",
+            ] and b_scale_category in ["scalar", "vector"]
+            is_block_dequant = (
+                a_scale_category == "matrix" and b_scale_category == "matrix"
+            )
+
+            if not (is_scalar_or_vector_dequant or is_block_dequant):
+                continue
+
+            if is_block_dequant and (bias is not None or M % 4 != 0):
+                continue
+
+            param = {
+                "M": M,
+                "N": N,
+                "K": K,
+                "a_scale_category": a_scale_category,
+                "b_scale_category": b_scale_category,
+                "use_bias": bias,
+                "in_dtype": in_dtype,
+                "out_dtype": out_dtype,
+            }
+            all_params.append(param)
+
+        return cls._rand_sample(all_params)
+
+    @staticmethod
+    def get_scale_shape(M, N, K, category, is_a_scale=True):
+        if category == "scalar":
+            return (1,)
+        if category == "vector":
+            if is_a_scale:
+                return (M,)
+            return (N,)
+        # a matrix
+        if is_a_scale:
+            return (M, ceil(K / 128))
+        return (ceil(K / 128), ceil(N / 128))
+
+    @staticmethod
+    def baseline_scaled_mm(
+        a: torch.Tensor,
+        b: torch.Tensor,
+        scale_a: torch.Tensor,
+        scale_b: torch.Tensor,
+        out_dtype: torch.dtype,
+        bias: Optional[torch.Tensor] = None,
+    ):
+        def group_broadcast(t: torch.Tensor, shape):
+            for i, s in enumerate(shape):
+                if t.shape[i] != s and t.shape[i] != 1:
+                    assert s % t.shape[i] == 0
+                    t = (
+                        t.unsqueeze(i + 1)
+                        .expand(*t.shape[: i + 1], s // t.shape[i], *t.shape[i + 1 :])
+                        .flatten(i, i + 1)
+                    )
+            return t
+
+        scale_a_full = group_broadcast(scale_a, a.shape)
+        scale_b_full = group_broadcast(scale_b, b.shape)
+
+        a_f32 = a.to(torch.float32)
+        b_f32 = b.to(torch.float32)
+
+        lhs = scale_a_full * a_f32
+        rhs = scale_b_full * b_f32
+
+        output = torch.mm(lhs, rhs).to(out_dtype)
+
+        if bias is not None:
+            output = output + bias
+
+        return output
+
+
+@pytest.mark.cutlass_scaled_mm
+@pytest.mark.skipif(
+    not (VLLM_AVAILABLE and CUDA_AVAILABLE),
+    reason="requires vLLM and NVIDIA Hopper architecture",
+)
+@pytest.mark.parametrize("p", CutlassScaledMMTestKit.get_test_params())
+def test_cutlass_scaled_mm(p):
+    kit = CutlassScaledMMTestKit
+
+    M, N, K = p["M"], p["N"], p["K"]
+    in_dtype = p["in_dtype"]
+    out_dtype = p["out_dtype"]
+    a_scale_category = p["a_scale_category"]
+    b_scale_category = p["b_scale_category"]
+
+    if in_dtype == torch.int8:
+        a = to_int8(torch.randn((M, K), device=flag_gems.device))
+        b = to_int8(
+            torch.randn((K, N), device=flag_gems.device).t().contiguous().t() * 5
+        )
+    else:
+        a = to_fp8(torch.randn((M, K), device=flag_gems.device))
+        b = to_fp8(torch.randn((K, N), device=flag_gems.device).t().contiguous().t())
+
+    a_scale_shape = kit.get_scale_shape(M, N, K, a_scale_category)
+    b_scale_shape = kit.get_scale_shape(M, N, K, b_scale_category, False)
+
+    scale_a = torch.randn(a_scale_shape, device=flag_gems.device, dtype=torch.float32)
+    scale_b = torch.randn(b_scale_shape, device=flag_gems.device, dtype=torch.float32)
+
+    scale_a = scale_a.contiguous()
+    # convert scale_b to col-major
+    # (for scalar/vector scale_b, this's a identical transformation)
+    scale_b = scale_b.t().contiguous().t()
+
+    bias = None
+    if p["use_bias"]:
+        bias = torch.randn((N,), device=flag_gems.device, dtype=out_dtype)
+
+    c = torch.empty((M, N), device=flag_gems.device, dtype=out_dtype)
+
+    flag_gems.cutlass_scaled_mm(c, a, b, scale_a, scale_b, bias)
+
+    output_ref = kit.baseline_scaled_mm(
+        a, b, scale_a.view(-1, 1), scale_b.view(1, -1), out_dtype, bias
+    )
+
+    if in_dtype == torch.int8:
+        rtol, atol = 1e-1, 1.0
+    else:
+        rtol, atol = 5e-1, 1.5e-1
+
+    torch.testing.assert_close(c, output_ref, rtol=rtol, atol=atol)
